+++ src/test/kotlin/org/pcsoft/framework/kunit/distance/KDistanceUnitSystemTest.kt
@@ -41,7 +41,7 @@ class KDistanceUnitSystemTest {
         fathoms to KDistanceUnit.FATHOM.baseValue,
         chains to KDistanceUnit.CHAIN.baseValue,
         furlongs to KDistanceUnit.FURLONG.baseValue,
-        lightYears to KDistanceUnit.LIGHT_YEAR.baseValue,
+        light.years to KDistanceUnit.LIGHT_YEAR.baseValue,
         parsecs to KDistanceUnit.PARSEC.baseValue
     )