+++ src/test/kotlin/org/pcsoft/framework/kunit/distance/KDistanceUnitTest.kt
@@ -40,7 +40,7 @@ class KDistanceUnitTest {
         fathoms to KDistanceUnit.FATHOM.baseValue,
         chains to KDistanceUnit.CHAIN.baseValue,
         furlongs to KDistanceUnit.FURLONG.baseValue,
-        lightYears to KDistanceUnit.LIGHT_YEAR.baseValue,
+        light.years to KDistanceUnit.LIGHT_YEAR.baseValue,
         parsecs to KDistanceUnit.PARSEC.baseValue,
         cubits to KDistanceUnit.CUBIT.baseValue,
         romanFeet to KDistanceUnit.ROMAN_FOOT.baseValue,